requirements.txt
python_version == "3.9.6"
# python_version == "3.11.9"

matplotlib==3.8.2
numpy==1.24.1
opencv_contrib_python==4.8.1.78
opencv_python==4.8.1.78
Pillow==9.3.0
# Pillow==10.4.0
gradio==4.44.0
keras==2.15.0
jellyfish==1.1.0
keras==2.15.0
scikit_learn==1.3.2
seaborn==0.13.2
tensorflow==2.15.0


# absl-py==2.1.0
# aiofiles==23.2.1
# annotated-types==0.7.0
# anyio==4.6.0
# appnope==0.1.4
# asttokens==2.4.1
# astunparse==1.6.3
# cachetools==5.5.0
# certifi==2024.8.30
# charset-normalizer==3.3.2
# click==8.1.7
# comm==0.2.2
# contourpy==1.3.0
# cycler==0.12.1
# debugpy==1.8.5
# decorator==5.1.1
# exceptiongroup==1.2.2
# executing==2.1.0
# fastapi==0.115.0
# ffmpy==0.4.0
# filelock==3.16.1
# flatbuffers==24.3.25
# fonttools==4.54.0
# fsspec==2024.9.0
# gast==0.6.0
# google-auth==2.35.0
# google-auth-oauthlib==1.2.1
# google-pasta==0.2.0
# gradio==4.44.0
# gradio_client==1.3.0
# grpcio==1.66.1
# h11==0.14.0
# h5py==3.11.0
# httpcore==1.0.5
# httpx==0.27.2
# huggingface-hub==0.25.1
# idna==3.10
# importlib_metadata==8.5.0
# importlib_resources==6.4.5
# ipykernel==6.29.5
# ipython==8.18.1
# jedi==0.19.1
# jellyfish==1.1.0
# Jinja2==3.1.4
# joblib==1.4.2
# jupyter_client==8.6.3
# jupyter_core==5.7.2
# keras==2.15.0
# kiwisolver==1.4.7
# libclang==18.1.1
# Markdown==3.7
# markdown-it-py==3.0.0
# MarkupSafe==2.1.5
# matplotlib==3.8.2
# matplotlib-inline==0.1.7
# mdurl==0.1.2
# ml-dtypes==0.2.0
# nest-asyncio==1.6.0
# numpy==1.24.1
# oauthlib==3.2.2
# opencv-contrib-python==4.8.1.78
# opencv-python==4.8.1.78
# opt-einsum==3.3.0
# orjson==3.10.7
# packaging==24.1
# pandas==2.2.3
# parso==0.8.4
# pexpect==4.9.0
# Pillow==9.3.0
# platformdirs==4.3.6
# prompt_toolkit==3.0.47
# protobuf==4.25.5
# psutil==6.0.0
# ptyprocess==0.7.0
# pure_eval==0.2.3
# pyasn1==0.6.1
# pyasn1_modules==0.4.1
# pydantic==2.9.2
# pydantic_core==2.23.4
# pydub==0.25.1
# Pygments==2.18.0
# pyparsing==3.1.4
# python-dateutil==2.9.0.post0
# python-multipart==0.0.10
# pytz==2024.2
# PyYAML==6.0.2
# pyzmq==26.2.0
# requests==2.32.3
# requests-oauthlib==2.0.0
# rich==13.8.1
# rsa==4.9
# ruff==0.6.7
# scikit-learn==1.3.2
# scipy==1.13.1
# seaborn==0.13.2
# semantic-version==2.10.0
# shellingham==1.5.4
# six==1.16.0
# sniffio==1.3.1
# stack-data==0.6.3
# starlette==0.38.6
# tensorboard==2.15.2
# tensorboard-data-server==0.7.2
# tensorflow==2.15.0
# tensorflow-estimator==2.15.0
# tensorflow-io-gcs-filesystem==0.37.1
# tensorflow-macos==2.15.0
# termcolor==2.4.0
# threadpoolctl==3.5.0
# tomlkit==0.12.0
# tornado==6.4.1
# tqdm==4.66.5
# traitlets==5.14.3
# typer==0.12.5
# typing_extensions==4.12.2
# tzdata==2024.2
# urllib3==2.2.3
# uvicorn==0.30.6
# wcwidth==0.2.13
# websockets==12.0
# Werkzeug==3.0.4
# wrapt==1.14.1
# zipp==3.20.2
